fixed name fclean, size_t to int cast

+++ srcs/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: averheij <averheij@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/10/29 11:02:49 by averheij       #+#    #+#                */
-/*   Updated: 2019/10/29 11:25:13 by averheij      ########   odam.nl         */
+/*   Updated: 2019/10/29 15:57:04 by averheij      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int		ft_memcmp(const void *s1, const void *s2, size_t n)
 	int		i;
 
 	i = 0;
-	while (i < n)
+	while (i < (int)n)
 	{
 		if (*(unsigned char*)(s1 + i) != *(unsigned char*)(s2 + i))
 			return (*(unsigned char*)(s1 + i) - *(unsigned char*)(s2 + i));
