+++ srcs/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: averheij <averheij@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/10/29 13:53:58 by averheij       #+#    #+#                */
-/*   Updated: 2019/10/29 14:01:32 by averheij      ########   odam.nl         */
+/*   Updated: 2019/10/29 15:57:49 by averheij      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int		ft_strncmp(const char *s1, const char *s2, size_t n)
 	int		i;
 
 	i = 0;
-	while (i < n && s1[i] && s2[i])
+	while (i < (int)n && s1[i] && s2[i])
 	{
 		if (s1[i] != s2[i])
 			return ((unsigned char)s1[i] - (unsigned char)s2[i]);
